refactor: split models/scan.ts into the HTTP surface and the two models
845 lines holding the scan wire shapes, the fetch calls, and two
independent models — a 157-line list model and a 410-line detail model
that share nothing but the transport.

- scan-api.ts          wire shapes and the calls that produce them
- scan-list-model.ts   the dashboard list: paging, decision filter, actions
- scan-detail-model.ts one scan's workbench: report, selection, shares,
                       gate decisions, and the polling to a terminal status

scan.ts stays the import path for all 17 consumers. Declaration census
run: 37 names before, 37 after, none dropped or duplicated. verify, knip
and build clean.
